requirements.txt
Django==1.7.5
Markdown==2.6
django-bootstrap3==5.1.1
django-filter==0.9.2
djangorestframework==3.0.5
